+++ projects/ngx-disqus/src/public_api.ts
@@ -1,0 +1,3 @@
+export { DisqusModule } from './lib/disqus.module';
+export { DisqusComment, DisqusReady } from './lib/disqus.model';
+export { DisqusService } from './lib/disqus.service';